+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/advancements/SVOAdvancementsProvider.java
@@ -1,6 +1,7 @@
 package com.CuteNekoDragon.Core.common.datagen.advancements;
 
 import com.CuteNekoDragon.Core.common.datagen.advancements.tabs.Chapter1;
+import com.CuteNekoDragon.Core.common.datagen.advancements.tabs.RemoveTabs;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.ExistingFileHelper;
@@ -11,6 +12,6 @@ import java.util.concurrent.CompletableFuture;
 
 public class SVOAdvancementsProvider extends ForgeAdvancementProvider {
     public SVOAdvancementsProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries, ExistingFileHelper existingFileHelper) {
-        super(output, registries, existingFileHelper, List.of(new Chapter1()));
+        super(output, registries, existingFileHelper, List.of(new Chapter1(), new RemoveTabs()));
     }
 }