+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/advancements/tabs/RemoveTabs.java
@@ -1,0 +1,17 @@
+package com.CuteNekoDragon.Core.common.datagen.advancements.tabs;
+
+import net.minecraft.advancements.Advancement;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.common.data.ForgeAdvancementProvider;
+
+import java.util.function.Consumer;
+
+public class RemoveTabs implements ForgeAdvancementProvider.AdvancementGenerator {
+    @Override
+    public void generate(HolderLookup.Provider provider, Consumer<Advancement> consumer, ExistingFileHelper existingFileHelper) {
+        Advancement.Builder.advancement()
+                .save(consumer, ResourceLocation.parse("minecraft:story/root"), existingFileHelper);
+    }
+}